First implementation of 64 bits (#2)
* feat: Migrate from x86 32-bit to x86-64 architecture with Limine bootloader

- Migrated entire codebase from i686 (32-bit) to x86_64 (64-bit) architecture
- Replaced custom boot.s with Limine bootloader protocol for modern UEFI/BIOS support
- Updated Makefile with x86-64 compiler flags (-m64, -mcmodel=kernel, -mno-red-zone)
- Added Limine bootloader integration with automatic download and ISO generation
- Implemented SYSCALL/SYSRET fast system call mechanism with MSR configuration
- Added framebuffer console

* feat: Replace VGA text mode with framebuffer-only console and migrate kernel panic to klog

- Removed VGA text mode support from console.c, now exclusively uses framebuffer console
- Simplified console.c to wrapper functions that forward all calls to fb_console
- Migrated kernel panic exception handler from direct VGA memory writes to KLOG_ERROR macros
- Changed QEMU display from virtio-vga to standard VGA (-vga std) in Makefile
- Removed VGA buffer management, scrolling, and hardware cursor code

* chore: Improve debugging and fix context switching RFLAGS handling

- Fixed RFLAGS save/restore order in switch_context/switch_task to prevent TF/IF flag leakage between threads
- Added Debug Exception (INT 0x01) handler in idt.c to clear Trap Flag and handle hardware breakpoints
- Added clear_debug_registers() function in kernel.c to disable hardware breakpoints at boot
- Fixed ELF loader to use HHDM virtual addresses from pmm_alloc_block() and zero pages via memset
- Fixed process_exec_and_wait to

* chore: Update .gitignore with absolute paths and remove duplicate entry

- Changed relative paths to absolute paths for lux-os/, limine/, and iso_root/ directories
- Removed duplicate alos.elf entry at end of file
- Added trailing newline to .gitignore

* chore: Revert to flat binary build and remove Limine bootloader artifacts

- Removed limine/, iso_root/, alos.iso, and alos.elf from .gitignore (now tracked)
- Deleted alos.iso and entire iso_root/ directory with Limine bootloader files
- Removed Limine UEFI boot files (BOOTIA32.EFI, BOOTX64.EFI) and BIOS boot files
- Deleted Limine configuration file (limine.conf)
- Updated alos.bin and alos.elf binaries

* feat: Add condition variable-based accept() blocking and fix thread sleep wakeup

- Added scheduler_wake_sleeping() call in timer_handler_c to wake threads whose sleep has expired
- Replaced busy-wait polling in sys_accept with condvar_wait on accept_cv/accept_mutex
- Added accept_cv condition variable and accept_mutex to tcp_socket_t for LISTEN sockets
- Modified tcp_handle_packet to broadcast accept_cv when client transitions to ESTABLISHED state
- Initialized accept_cv and accept_mutex in tcp_init_

* feat: Increase ARP cache size to 1024 entries and improve cache logging

- Increased ARP_CACHE_SIZE from 16 to 1024 entries to support larger networks
- Changed arp_cache_add() logging from KLOG_INFO to KLOG_DEBUG_DEC with cache index display
- Removed trailing whitespace in arp_cache_add() function

* feat: Add IRQ-safe spinlock variants to prevent scheduler deadlocks in interrupt context

- Added spinlock_irqsave() and spinlock_irqrestore() functions to disable/restore interrupts during lock acquisition
- Replaced spinlock_lock/unlock with IRQ-safe variants in scheduler_tick(), scheduler_enqueue(), scheduler_dequeue(), and check_thread_timeouts()
- Added comments explaining IRQ-safe locking is needed when scheduler functions are called from interrupt context (e.g., timer IRQ, tcp_handle_packet)
-

* feat: Fix syscall/interrupt handling and improve keyboard driver with semaphore-based blocking

- Fixed INT 0x80 handler to push dummy int_no/error_code before registers to match syscall_regs_t structure layout
- Fixed syscall_entry SYSRET sequence to use temporary register (r10) instead of "pop rsp" which corrupts stack pointer
- Added comprehensive debug logging to Debug Exception handler (INT 0x01) showing RIP/RSP/CS/SS registers
- Migrated keyboard driver from busy-wait polling to semaphore-based blocking in

* docs: Reorganize README with feature checklist and expand roadmap to 150+ items

- Converted flat feature list to categorized checklist format with [x] completed and [ ] planned items
- Expanded roadmap from ~40 items to 150+ items across 15 categories (Memory, IPC, Storage, Network, Drivers, UI, Security, etc.)
- Moved Linux compatibility documentation reference before License section
- Removed duplicate networking SLIRP configuration section (was listed twice)
- Reordered sections: Features →

+++ src/kernel/keymap.c
@@ -133,7 +133,7 @@ static const keymap_t* current_keymap = &keymap_qwerty_us;
 
 /* Table des keymaps enregistrées */
 static const keymap_t* registered_keymaps[MAX_KEYMAPS];
-static int keymap_count = 0;
+static size_t keymap_count = 0;
 
 /* ========================================
  * Implémentation des fonctions
@@ -182,7 +182,7 @@ const keymap_t* keymap_find_by_name(const char* name)
         return NULL;
     }
     
-    for (int i = 0; i < keymap_count; i++) {
+    for (size_t i = 0; i < keymap_count; i++) {
         if (strcmp(registered_keymaps[i]->name, name) == 0) {
             return registered_keymaps[i];
         }
@@ -194,7 +194,7 @@ const keymap_t* keymap_find_by_name(const char* name)
 /**
  * Récupère la liste de toutes les keymaps disponibles.
  */
-const keymap_t** keymap_list_all(int* count)
+const keymap_t** keymap_list_all(size_t* count)
 {
     if (count != NULL) {
         *count = keymap_count;
@@ -212,7 +212,7 @@ bool keymap_register(const keymap_t* map)
     }
     
     /* Vérifier qu'elle n'est pas déjà enregistrée */
-    for (int i = 0; i < keymap_count; i++) {
+    for (size_t i = 0; i < keymap_count; i++) {
         if (registered_keymaps[i] == map) {
             return true;  /* Déjà enregistrée */
         }
